requirements.txt
amqp==2.3.2
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.5
coverage==4.5.2
funcsigs==1.0.2
kombu==4.2.1
more-itertools==4.3.0
pathlib2==2.3.3
pluggy==0.8.0
py==1.7.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
pytz==2018.7
scandir==1.9.0
six==1.11.0
vine==1.1.4
ipaddress=1.0.22
python-dateutil=2.7.5
text-unidecode=1.2
Faker=1.0.0
factory-boy=2.11.1
mock=2.0.0
pbr=5.1.1

celery==4.2.1

Django==1.11.17
djangorestframework==3.9.0
django-filter==1.1.0